scripts/ai_agent/pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9.0,<4.0"
dependencies = [
    "langchain-community (>=0.3.15,<0.4.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "langgraph (>=0.2.66,<0.3.0)",
    "mysql-connector-python (>=9.2.0,<10.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "langchain-openai (>=0.3.1,<0.4.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
